avance 09-07-2025

+++ src/main/java/com/transporte_gomez/erp/repository/DevolucionDetalleRepository.java
@@ -3,5 +3,5 @@ package com.transporte_gomez.erp.repository;
 import com.transporte_gomez.erp.entity.DevolucionDetalleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DevolucionDetalleRepository extends JpaRepository<DevolucionDetalleEntity, Integer> {
+public interface DevolucionDetalleRepository extends JpaRepository<DevolucionDetalleEntity, Long> {
 }
